eaton-mib.h renamed into separate constituent headers
