--- drivers/eaton-mib.h
@@ -1,13 +1,0 @@
-#ifndef EATON_MIB_H
-#define EATON_MIB_H
-
-#include "main.h"
-#include "snmp-ups.h"
-
-extern mib2nut_info_t	aphel_genesisII;
-extern mib2nut_info_t	aphel_revelation;
-extern mib2nut_info_t	eaton_marlin;
-extern mib2nut_info_t	pulizzi_switched1;
-extern mib2nut_info_t	pulizzi_switched2;
-
-#endif /* EATON_MIB_H */
