+++ drivers/eaton-pdu-marlin-mib.h
@@ -1,0 +1,32 @@
+/* eaton-pdu-marlin-mib.h - subdriver to monitor Eaton ePDU SNMP devices with NUT
+ *
+ *  Copyright (C)
+ *  2010        Arjen de Korte <adkorte-guest@alioth.debian.org>
+ *  2011 - 2012 Arnaud Quette <arnaud.quette@free.fr>
+ *  2017        Arnaud Quette <ArnaudQuette@eaton.com>
+ *  2017        Jim Klimov <EvgenyKlimov@eaton.com>
+ *
+ *  This program is free software; you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation; either version 2 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program; if not, write to the Free Software
+ *  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
+ */
+
+#ifndef EATON_EPDU_MARLIN_MIB_H
+#define EATON_EPDU_MARLIN_MIB_H
+
+#include "main.h"
+#include "snmp-ups.h"
+
+extern mib2nut_info_t	eaton_marlin;
+
+#endif /* EATON_EPDU_MARLIN_MIB_H */
