+++ drivers/snmp-ups.c
@@ -44,7 +44,10 @@
 #include "mge-mib.h"
 #include "netvision-mib.h"
 #include "powerware-mib.h"
-#include "eaton-mib.h"
+#include "eaton-pdu-genesis2-mib.h"
+#include "eaton-pdu-marlin-mib.h"
+#include "eaton-pdu-pulizzi-mib.h"
+#include "eaton-pdu-revelation-mib.h"
 #include "raritan-pdu-mib.h"
 #include "raritan-px2-mib.h"
 #include "baytech-mib.h"
